+++ egui_commonmark_backend/src/lib.rs
@@ -26,6 +26,9 @@ pub use {
 // The only struct that is allowed to use directly. (If one does not need egui_commonmark)
 pub use misc::CommonMarkCache;
 
+#[cfg(feature = "better_syntax_highlighting")]
+pub use syntect;
+
 /// Takes [`egui::Ui`], the math text to be rendered and whether it is inline
 pub type RenderMathFn = dyn Fn(&mut egui::Ui, &str, bool);
 /// Takes [`egui::Ui`] and the html text to be rendered/used